requirements.txt
alembic==1.4.2
attrs==19.3.0
blinker==1.4
click==7.1.2
flasgger==0.9.4
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.3.0
mysqlclient==1.4.6
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
PyYAML==5.3.1
redis==3.5.2
six==1.14.0
SQLAlchemy==1.3.17
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
